+++ docs/src/reference/cli/changelog.rst
@@ -1,6 +1,32 @@
 CLI Changelog
 =============
 
+.. _cli-21.12.0:
+
+21.12.0 (2021-12-14)
+--------------------
+
+.. start-21.12.0
+
+Added
+*****
+
+* Include more GCP roles when verifying a job's service account in ``klio job verify``.
+
+Fixed
+*****
+
+* Correctly validate existence of Dataflow-related Klio config when running on Dataflow (and not just "not --direct-runner").
+* Print out ``gcloud`` commands to add necessary GCP roles when running ``klio job verify --create-resources``.
+
+Changed
+*******
+
+* When running a job, effective config is no longer written to ``klio-job-run-effective.yaml``, but instead to a temp file.  This file no longer needs to be included in ``setup.py`` projects (See `PR 233 <https://github.com/spotify/klio/pull/233>`_).
+* Error out when a user tries to run a Dataflow-based job with Stackdriver log-based metrics client configured.
+
+.. end-21.12.0
+
 .. _cli-21.9.0:
 
 21.9.0 (2021-10-12)